// frontend/web/src/lib/shared/enums/category.enum.ts
export enum ECategory {
  Estrategia = "Estratégia",
  Cooperativos = "Cooperativos",
  Deducao = "Dedução",
  PartyGame = "Party Game",
  BlefeEnganacao = "Blefe e Enganação",
  Destreza = "Destreza",
  JogosAbstratos = "Jogos Abstratos",
  LogicaQuebraCabeca = "Lógica e Quebra-Cabeça",
  AventuraTematicos = "Aventura e Temáticos",
  Negociacao = "Negociação",
  Outros = "Outros",
}